+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to **Pipecat** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.8] - 2025-09-18
+
+## Changed
+
+- We now just write one single file instead of one per client session.
+
+  ```python
+  whisker = WhiskerObserver(pipeline, file_name="whisker.bin")
+  ```
+
+## Fixed
+
+- Fixed a client issue that was causing the pipeline to be reset when receving
+  messages.
+
 ## [0.0.7] - 2025-09-18
 
 ## Added